HILDA dictionary in a data.table format.

A data.table that contains a dictionary of the loaded HILDA datasets.

data(hil_dict)

Format

A data table with many rows and 3 variables:

var

(character()) variable names

wave

(list(integer())) waves that the variable was recorded

label

(character()) short description of the variable

Examples

hil_dict

#>            var            wave
#>     1: xwaveid 1,2,3,4,5,6,...
#>     2:    hhid 1,2,3,4,5,6,...
#>     3:   hhpno 1,2,3,4,5,6,...
#>     4:   hhpid 1,2,3,4,5,6,...
#>     5: hhstate 1,2,3,4,5,6,...
#>    ---                        
#> 10279: fimfeue              16
#> 10280: fimfbss              16
#> 10281: fimfbfi              16
#> 10282: fimfd2d              16
#> 10283:  edasch              16
#>                                                                                   label
#>     1:                                                                 XW Cross wave ID
#>     2:                                                                  HF Household ID
#>     3:                                                                 HF Person number
#>     4:                                                                     HF Person ID
#>     5:                                                                         HF State
#>    ---                                                                                 
#> 10279: SCQ:C7f Managing finances - I always make sure I have money saved up for emergen
#> 10280: SCQ:C7g Managing finances - I do a good job of balancing my spending and savings
#> 10281: SCQ:C7h Managing finances - I feel very comfortable dealing with banks and other
#> 10282: SCQ:C7i Managing finances - I am good at dealing with day-to-day financial matte
#> 10283:                        CPQ:A2a At school at previous interview fed forward check